Added registering new devices and pop up messages
New devices can connect to the TCP server, which then
sends the IP address and MAC of the device to the web
server. The web server sends a response and the AJAX of
the main page updates with a pop up box. Notifying the
user that a new MSP430 is now online.