Created firmware project, can register with server
Created new project for MSP430 server. Imported the CC3000 library and
the 1.60 release of MSP430 Driver Library. Currently the MSP430 can
connect to the backend server and register.

Had to make a slight modification to the CC3000 stack, had to remove the
waiting for response after sending a message over TCP. Wireshark shows
that the 430 gets the ACK but the 430 seems to be waiting for an event.